requirements.txt
streamlit
pandas
numpy
plotly
python-dotenv
sentence-transformers
faiss-cpu
PyPDF2
anthropic